embedding-balancer/requirements.txt
# Listed dependencies as of 2020-07-25
docutils==0.15.2
falcon==2.0.0
gunicorn==20.0.4
#  - setuptools [required: >=3.0, installed: 46.1.3]
numpy==1.18.3
PyYAML==5.3.1
requests==2.24.0
